Updating the changes for scripts

+++ utilities/common/financial_utils.py
@@ -1,6 +1,7 @@
 from datetime import date
 from decimal import Decimal, ROUND_HALF_UP
 
+
 class FinancialCalculator:
     @staticmethod
     def calculate_roi(current_value: float, initial_investment: float) -> float:
@@ -10,13 +11,14 @@ class FinancialCalculator:
         return ((current_value - initial_investment) / initial_investment) * 100
 
     @staticmethod
-    def calculate_xirr(current_value: float, initial_investment: float, investment_date: date, current_date: date) -> float:
+    def calculate_xirr(current_value: float, initial_investment: float, investment_date: date,
+                       current_date: date) -> float:
         """
         Calculates a simplified XIRR (Annualized Return).
         """
         if initial_investment == 0:
             return 0.0
-            
+
         days_invested = (current_date - investment_date).days
         years = days_invested / 365.0
 
@@ -25,8 +27,8 @@ class FinancialCalculator:
 
         if years < 1:
             return ((current_value - initial_investment) / initial_investment) * 100
-            
-        return (((current_value / initial_investment) ** (1 / years)) - 1) * 100
+
+        return (((current_value / initial_investment) ** Decimal((1 / years))) - 1) * 100
 
     @staticmethod
     def round2(x) -> float:
